+++ DeviceControl.CoreTests/DAL/DataAccessUtils.cs
@@ -13,7 +13,7 @@ namespace DeviceControl.CoreTests.DAL
 
         static DataAccessUtils()
         {
-            AppSettings = GetAppSettings(EnumDb.Develop);
+            AppSettings = GetAppSettings(EnumDb.Debug);
             DataAccess = GetDataAccess();
         }
 
@@ -23,8 +23,8 @@ namespace DeviceControl.CoreTests.DAL
                 return AppSettings;
             AppSettings = db switch
             {
-                EnumDb.Develop => new CoreSettingsEntity("DEV1C.kolbasa-vs.local\\DVLP", "ScalesDB", false, "scale01", "scale01"),
-                EnumDb.Product => new CoreSettingsEntity("192.168.0.26\\WMS", "ScalesDB", false, "scale01", "scale01"),
+                EnumDb.Debug => new CoreSettingsEntity("CREATIO\\INS1", "SCALES", false, "scale01", "scale01"),
+                EnumDb.Release => new CoreSettingsEntity("PALYCH\\LUTON", "ScalesDB", false, "scale01", "scale01"),
                 _ => null
             };
             AppSettings?.CheckProperties();