+++ DeviceControl/CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.290] - 2021-07-26
 ### Changed
-- DB location
+- Debug DB location
 
 ## [0.2.280] - 2021-07-20
 ### Added